Hammer-2 : updated as per review comments

+++ src/components/ManageApp/Member.tsx
@@ -2,11 +2,13 @@ import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { deleteRoles } from '../../actions/OaRoleActions';
 import './member.scss';
+import { deleteAppSpace } from '../../actions/AppSpaceAction';
 
 interface Props {
   member: any;
   domainId: string;
-  domainType: string;
+  domainType?: string;
+  owner: string;
 }
 
 const Member = (props: Props) => {
@@ -15,20 +17,42 @@ const Member = (props: Props) => {
   const [confirmPromptOpen, setConfirmPromptOpen] = useState(false);
 
   const remove = () => {
-    dispatch(
-      deleteRoles(
-        authorization,
-        props.domainType,
-        props.member._id,
-        props.domainId
-      )
-    );
+    if (
+      props.member.email === props.owner ||
+      props.member._id === props.owner
+    ) {
+      return;
+    }
+    if (props.domainType) {
+      dispatch(
+        deleteRoles(
+          authorization,
+          props.domainType,
+          props.member._id,
+          props.domainId
+        )
+      );
+    } else {
+      dispatch(
+        deleteAppSpace(authorization, props.member.spaceId, props.domainId)
+      );
+    }
   };
 
   return (
     <div className="member" id={props.member._id}>
-      <div className="title typography-6">{props.member.email}</div>
-      <div className="typography-6">{`${props.member.lastName}, ${props.member.firstName}`}</div>
+      {props.domainType && (
+        <>
+          <div className="title typography-6">{props.member.email}</div>
+          <div className="typography-6">{`${props.member.lastName}, ${props.member.firstName}`}</div>
+        </>
+      )}
+      {!props.domainType && (
+        <>
+          <div className="title typography-6">{props.member.spaceId}</div>
+          <div className="typography-6">{props.member.name}</div>
+        </>
+      )}
       <div className="action">
         {!confirmPromptOpen && (
           <div
